UI Optimizations
Added a title and a simple icon for the application. Also cleaned up the ui by removing the ticket number field as it was a dummy field anyway.

Also removed the start and end date time labels in the tracking controls, as for now I don't see any benefit of having them around.

The reset button was also removed as the associated functionality is a responsibility of the start / stop button and should not be done by an additional control.