Implemented various time tracking features
Added support for displaying the current work week with all its time tracking entries grouped by day and project (on that day)

Also implemented support for deleting time entries and updating their duration. If a duration gets updated, the end time of the entry is adjusted. The duration itself is calculated based on the start and end of a time entry and is never persisted.

Also implemented a different location of the database file. This is now depending on the app mode. If the app mode is dev the database file will reside in the wd, whereas it will be placed in the user's home directory if the app mode is prod. This can be toggled using the 'dev' command line switch.

Also implemented a command line parameter 'customTime' that can be used to set the
time used by the application as today to 12.11.21.
This is the date where the test data is located at. The test data can be generated / recreated using the testData command line flag.