Removed module structure
As track never really used a server and was primarily used as a
standalon offline application, the server module has been removed.
Along with this, also the core module has been removed, as it doesn't
make sense to keep possible shared code, that no longer shared as the
server has been removed, in a separate module.

For simplicity the gui module was removed too and the project was
transformed to a single root module project.