Remaining hours to work
Implementing a bar to show how many hours one needs to work this week in order to get to the desired total hours. The standard for work week hours is set to 40, but can be customized using the settings.

Every tracking entry of a given week will be subtracted from the configured amount of hours to work per week.

There is still some technical stuff to be done like adapting the database types for storing the total hours of work per week. Those are stored as int but should be long, and also it should be clearly defined that we are storing seconds here not the hours only.